feat: refactor TypeScript code generation with new modules and functions
- Removed the `typescript.ts` file and reorganized the codebase into multiple modules for better structure and maintainability.
- Introduced new files: `augmentation.ts`, `control.ts`, `enum.ts`, `function.ts`, `type.ts`, and `index.ts` to encapsulate related functionalities.
- Added functions for generating TypeScript augmentations, control flow statements, enums, and function declarations, enhancing the overall code generation capabilities.
- Updated tests to cover new functionalities and ensure correctness across the newly structured modules.
