+++ src/typescript/index.ts
@@ -1,0 +1,35 @@
+export type {
+  TypeGeneric,
+  TypeField,
+  FunctionOpts,
+  TypeObjectWithJSDoc,
+  TypeObject,
+  TypeObjectField,
+  GenInterfaceOptions,
+  EnumMemberValue,
+  GenEnumOptions,
+  GenTypeAliasOptions,
+  GenVariableOptions,
+  GenIfOptions,
+  GenElseOptions,
+  GenTryOptions,
+  GenCatchOptions,
+} from "./types";
+
+export {
+  genTypeAlias,
+  genVariable,
+  genTypeExport,
+  genInlineTypeImport,
+  genTypeObject,
+  genProperty,
+  genInterface,
+} from "./type";
+
+export { genEnum, genConstEnum } from "./enum";
+
+export { genParam, genFunction, genBlock } from "./function";
+
+export { genAugmentation, genDeclareNamespace } from "./augmentation";
+
+export { genIf, genElseIf, genElse, genTry, genCatch } from "./control";
